+++ Assets/Scripts/Enemies/EnemyFall.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 20b107c4624e2b3479830e1e4b1513e6
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
